only delete bin folder in static builds

+++ ports/opencv4/portfile.cmake
@@ -558,8 +558,9 @@ find_dependency(Qt${USE_QT_VERSION} COMPONENTS OpenGL)")
   endif()
 
   file(WRITE "${CURRENT_PACKAGES_DIR}/share/opencv4/OpenCVModules.cmake" "${OPENCV_MODULES}")
+endif()
 
-
+if(VCPKG_LIBRARY_LINKAGE STREQUAL "static)
   file(REMOVE_RECURSE "${CURRENT_PACKAGES_DIR}/bin" "${CURRENT_PACKAGES_DIR}/debug/bin")
 endif()
